requirements.txt
Django
djangorestframework
django-filter
django-cors-headers
django-import-export
djangorestframework-simplejwt
dj-rest-auth
gunicorn
Pillow